requirements.txt
django==3.1.2
requests==2.24.0
python-dateutil==2.8.1
captcha==0.3